Сделал предварительное разделение на модули.
